requirements.txt
# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9  # for file uploads

# Vision-Language model
transformers>=4.40.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.3.0

# Camera input
opencv-python>=4.9.0

# LLM fallback
openai>=1.30.0

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0
pandas>=2.2.0
pyarrow>=16.0.0
scikit-learn>=1.5.0
datasets>=2.19.0
huggingface_hub>=0.23.0
sentence-transformers>=3.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
umap-learn>=0.5.6

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # for FastAPI test client
